+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file, following t
 ### Added
 - traverse atoms before group-bonds in `MMTF.traverse`
 - check left-alignment of chainIds & chainNames
+- param argument for MMTF.traverse to traverse only the first model
 
 ### Changed
 - fix chainIds & chainNames to be left-alignment in test data